Speech: allow `auto` setting for language to derive from display language (fix #206321)
