+++ src/vs/workbench/contrib/accessibility/browser/accessibilityConfiguration.ts
@@ -9,7 +9,7 @@ import { Registry } from 'vs/platform/registry/common/platform';
 import { RawContextKey } from 'vs/platform/contextkey/common/contextkey';
 import { workbenchConfigurationNodeBase, Extensions as WorkbenchExtensions, IConfigurationMigrationRegistry, ConfigurationKeyValuePairs } from 'vs/workbench/common/configuration';
 import { AccessibilityAlertSettingId, AccessibilitySignal } from 'vs/platform/accessibilitySignal/browser/accessibilitySignalService';
-import { ISpeechService } from 'vs/workbench/contrib/speech/common/speechService';
+import { ISpeechService, SPEECH_LANGUAGES, SPEECH_LANGUAGE_CONFIG } from 'vs/workbench/contrib/speech/common/speechService';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { IWorkbenchContribution } from 'vs/workbench/common/contributions';
 import { Event } from 'vs/base/common/event';
@@ -664,10 +664,9 @@ export function registerAccessibilityConfiguration() {
 
 export const enum AccessibilityVoiceSettingId {
 	SpeechTimeout = 'accessibility.voice.speechTimeout',
-	SpeechLanguage = 'accessibility.voice.speechLanguage'
+	SpeechLanguage = SPEECH_LANGUAGE_CONFIG
 }
 export const SpeechTimeoutDefault = 1200;
-const SpeechLanguageDefault = 'en-US';
 
 export class DynamicSpeechAccessibilityConfiguration extends Disposable implements IWorkbenchContribution {
 
@@ -703,10 +702,10 @@ export class DynamicSpeechAccessibilityConfiguration extends Disposable implemen
 					'tags': ['accessibility']
 				},
 				[AccessibilityVoiceSettingId.SpeechLanguage]: {
-					'markdownDescription': localize('voice.speechLanguage', "The language that voice speech recognition should recognize."),
+					'markdownDescription': localize('voice.speechLanguage', "The language that voice speech recognition should recognize. Select `auto` to use VS Code's configured display language."),
 					'type': 'string',
 					'enum': languagesSorted,
-					'default': SpeechLanguageDefault,
+					'default': 'auto',
 					'tags': ['accessibility'],
 					'enumDescriptions': languagesSorted.map(key => languages[key].name),
 					'enumItemLabels': languagesSorted.map(key => languages[key].name)
@@ -717,84 +716,10 @@ export class DynamicSpeechAccessibilityConfiguration extends Disposable implemen
 
 	private getLanguages(): { [locale: string]: { name: string } } {
 		return {
-			['da-DK']: {
-				name: localize('speechLanguage.da-DK', "Danish (Denmark)")
+			['auto']: {
+				name: localize('speechLanguage.auto', "Auto (Use Display Language)")
 			},
-			['de-DE']: {
-				name: localize('speechLanguage.de-DE', "German (Germany)")
-			},
-			['en-AU']: {
-				name: localize('speechLanguage.en-AU', "English (Australia)")
-			},
-			['en-CA']: {
-				name: localize('speechLanguage.en-CA', "English (Canada)")
-			},
-			['en-GB']: {
-				name: localize('speechLanguage.en-GB', "English (United Kingdom)")
-			},
-			['en-IE']: {
-				name: localize('speechLanguage.en-IE', "English (Ireland)")
-			},
-			['en-IN']: {
-				name: localize('speechLanguage.en-IN', "English (India)")
-			},
-			['en-NZ']: {
-				name: localize('speechLanguage.en-NZ', "English (New Zealand)")
-			},
-			[SpeechLanguageDefault]: {
-				name: localize('speechLanguage.en-US', "English (United States)")
-			},
-			['es-ES']: {
-				name: localize('speechLanguage.es-ES', "Spanish (Spain)")
-			},
-			['es-MX']: {
-				name: localize('speechLanguage.es-MX', "Spanish (Mexico)")
-			},
-			['fr-CA']: {
-				name: localize('speechLanguage.fr-CA', "French (Canada)")
-			},
-			['fr-FR']: {
-				name: localize('speechLanguage.fr-FR', "French (France)")
-			},
-			['hi-IN']: {
-				name: localize('speechLanguage.hi-IN', "Hindi (India)")
-			},
-			['it-IT']: {
-				name: localize('speechLanguage.it-IT', "Italian (Italy)")
-			},
-			['ja-JP']: {
-				name: localize('speechLanguage.ja-JP', "Japanese (Japan)")
-			},
-			['ko-KR']: {
-				name: localize('speechLanguage.ko-KR', "Korean (South Korea)")
-			},
-			['nl-NL']: {
-				name: localize('speechLanguage.nl-NL', "Dutch (Netherlands)")
-			},
-			['pt-PT']: {
-				name: localize('speechLanguage.pt-PT', "Portuguese (Portugal)")
-			},
-			['pt-BR']: {
-				name: localize('speechLanguage.pt-BR', "Portuguese (Brazil)")
-			},
-			['ru-RU']: {
-				name: localize('speechLanguage.ru-RU', "Russian (Russia)")
-			},
-			['sv-SE']: {
-				name: localize('speechLanguage.sv-SE', "Swedish (Sweden)")
-			},
-			['tr-TR']: {
-				name: localize('speechLanguage.tr-TR', "Turkish (Turkey)")
-			},
-			['zh-CN']: {
-				name: localize('speechLanguage.zh-CN', "Chinese (Simplified, China)")
-			},
-			['zh-HK']: {
-				name: localize('speechLanguage.zh-HK', "Chinese (Traditional, Hong Kong)")
-			},
-			['zh-TW']: {
-				name: localize('speechLanguage.zh-TW', "Chinese (Traditional, Taiwan)")
-			}
+			...SPEECH_LANGUAGES
 		};
 	}
 }
