+++ src/vs/workbench/contrib/speech/common/speechService.ts
@@ -10,6 +10,7 @@ import { IDisposable } from 'vs/base/common/lifecycle';
 import { RawContextKey } from 'vs/platform/contextkey/common/contextkey';
 import { ExtensionIdentifier } from 'vs/platform/extensions/common/extensions';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
+import { language } from 'vs/base/common/platform';
 
 export const ISpeechService = createDecorator<ISpeechService>('speechService');
 
@@ -96,4 +97,107 @@ export interface ISpeechService {
 	 * stopped.
 	 */
 	recognizeKeyword(token: CancellationToken): Promise<KeywordRecognitionStatus>;
+}
+
+export const SPEECH_LANGUAGE_CONFIG = 'accessibility.voice.speechLanguage';
+
+export const SPEECH_LANGUAGES = {
+	['da-DK']: {
+		name: localize('speechLanguage.da-DK', "Danish (Denmark)")
+	},
+	['de-DE']: {
+		name: localize('speechLanguage.de-DE', "German (Germany)")
+	},
+	['en-AU']: {
+		name: localize('speechLanguage.en-AU', "English (Australia)")
+	},
+	['en-CA']: {
+		name: localize('speechLanguage.en-CA', "English (Canada)")
+	},
+	['en-GB']: {
+		name: localize('speechLanguage.en-GB', "English (United Kingdom)")
+	},
+	['en-IE']: {
+		name: localize('speechLanguage.en-IE', "English (Ireland)")
+	},
+	['en-IN']: {
+		name: localize('speechLanguage.en-IN', "English (India)")
+	},
+	['en-NZ']: {
+		name: localize('speechLanguage.en-NZ', "English (New Zealand)")
+	},
+	['en-US']: {
+		name: localize('speechLanguage.en-US', "English (United States)")
+	},
+	['es-ES']: {
+		name: localize('speechLanguage.es-ES', "Spanish (Spain)")
+	},
+	['es-MX']: {
+		name: localize('speechLanguage.es-MX', "Spanish (Mexico)")
+	},
+	['fr-CA']: {
+		name: localize('speechLanguage.fr-CA', "French (Canada)")
+	},
+	['fr-FR']: {
+		name: localize('speechLanguage.fr-FR', "French (France)")
+	},
+	['hi-IN']: {
+		name: localize('speechLanguage.hi-IN', "Hindi (India)")
+	},
+	['it-IT']: {
+		name: localize('speechLanguage.it-IT', "Italian (Italy)")
+	},
+	['ja-JP']: {
+		name: localize('speechLanguage.ja-JP', "Japanese (Japan)")
+	},
+	['ko-KR']: {
+		name: localize('speechLanguage.ko-KR', "Korean (South Korea)")
+	},
+	['nl-NL']: {
+		name: localize('speechLanguage.nl-NL', "Dutch (Netherlands)")
+	},
+	['pt-PT']: {
+		name: localize('speechLanguage.pt-PT', "Portuguese (Portugal)")
+	},
+	['pt-BR']: {
+		name: localize('speechLanguage.pt-BR', "Portuguese (Brazil)")
+	},
+	['ru-RU']: {
+		name: localize('speechLanguage.ru-RU', "Russian (Russia)")
+	},
+	['sv-SE']: {
+		name: localize('speechLanguage.sv-SE', "Swedish (Sweden)")
+	},
+	['tr-TR']: {
+		name: localize('speechLanguage.tr-TR', "Turkish (Turkey)")
+	},
+	['zh-CN']: {
+		name: localize('speechLanguage.zh-CN', "Chinese (Simplified, China)")
+	},
+	['zh-HK']: {
+		name: localize('speechLanguage.zh-HK', "Chinese (Traditional, Hong Kong)")
+	},
+	['zh-TW']: {
+		name: localize('speechLanguage.zh-TW', "Chinese (Traditional, Taiwan)")
+	}
+};
+
+export function speechLanguageConfigToLanguage(config: unknown, lang = language): string {
+	if (typeof config === 'string') {
+		if (config === 'auto') {
+			if (lang !== 'en') {
+				const displayLanguage = lang;
+				const displayLanguageParts = displayLanguage.split('-');
+				const displayLocale = `${displayLanguageParts[0]}-${(displayLanguageParts[1] ?? displayLanguageParts[0]).toUpperCase()}`;
+
+				return speechLanguageConfigToLanguage(displayLocale);
+			}
+		} else {
+			if (SPEECH_LANGUAGES[config as keyof typeof SPEECH_LANGUAGES]) {
+				return config;
+			}
+		}
+	}
+
+	return 'en-US';
 }
